refactor(service): move resolve_input to workflows.input_resolution (ADR-061 Phase 4)
Extracts resolve_input and its zero-PolicyFile-dependency helpers
(detect_binary_format, sniff_text_format, collect_metadata,
load_env_matrix, and their private helpers) out of service.py into the
new abicheck.workflows.input_resolution, dropping service.py from 886
to 439 lines. abicheck.serialization joins modules.yaml's
public_root_surfaces so the new module can reach load_snapshot without
the deeper storage reclassification serialization.py's own debt entry
is still blocked on.

service.py re-exports the moved names via a plain static import;
compare_snapshots/load_suppression_and_policy stay behind since they
still need PolicyFile, whose classification remains this ADR's own
open question. Repointed tests that intercepted resolve_input's
internal run_dump/load_snapshot/detect_binary_format/sniff_text_format
calls at abicheck.workflows.input_resolution instead of abicheck.service,
since those bare-name calls now resolve against the new module's own
globals.
