+++ tests/test_cli_comparability_gate.py
@@ -20,7 +20,7 @@ dedicated ``except (ProfileMismatchError, ScopeMismatchError)`` branch in
 directly, bypassing ``CompareRequest``/``run_compare_request`` — see
 ``abicheck/comparability.py``'s module docstring). Mirrors
 ``tests/test_cli_coverage.py::TestCompareApiBreakExitCode``'s pattern of
-monkeypatching ``abicheck.service.load_snapshot``/``compare_snapshots``
+monkeypatching ``abicheck.workflows.input_resolution.load_snapshot``/``compare_snapshots``
 instead of driving a real dump, since the gate's own logic is already
 covered end-to-end by ``tests/test_checker_comparability_gate.py``."""
 
@@ -51,7 +51,7 @@ class TestNotComparableExitCode:
     def test_scope_mismatch_exits_16(self, tmp_path, monkeypatch):
         old_p, new_p = _write_placeholder_inputs(tmp_path)
         snap = AbiSnapshot(library="libfoo.so.1", version="1.0")
-        monkeypatch.setattr("abicheck.service.load_snapshot", lambda _: snap)
+        monkeypatch.setattr("abicheck.workflows.input_resolution.load_snapshot", lambda _: snap)
 
         def _raise(*_a, **_kw):
             raise ScopeMismatchError(
@@ -69,7 +69,7 @@ class TestNotComparableExitCode:
     def test_profile_mismatch_exits_16(self, tmp_path, monkeypatch):
         old_p, new_p = _write_placeholder_inputs(tmp_path)
         snap = AbiSnapshot(library="libfoo.so.1", version="1.0")
-        monkeypatch.setattr("abicheck.service.load_snapshot", lambda _: snap)
+        monkeypatch.setattr("abicheck.workflows.input_resolution.load_snapshot", lambda _: snap)
 
         def _raise(*_a, **_kw):
             raise ProfileMismatchError("profile_fingerprint mismatch: dep.h changed")
@@ -110,7 +110,7 @@ class TestNotComparableExitCode:
     def test_json_format_emits_verdict_null_with_reason(self, tmp_path, monkeypatch):
         old_p, new_p = _write_placeholder_inputs(tmp_path)
         snap = AbiSnapshot(library="libfoo.so.1", version="1.0")
-        monkeypatch.setattr("abicheck.service.load_snapshot", lambda _: snap)
+        monkeypatch.setattr("abicheck.workflows.input_resolution.load_snapshot", lambda _: snap)
 
         def _raise(*_a, **_kw):
             raise ScopeMismatchError("scope drift")
@@ -133,7 +133,7 @@ class TestNotComparableExitCode:
     def test_sarif_format_emits_failed_invocation(self, tmp_path, monkeypatch):
         old_p, new_p = _write_placeholder_inputs(tmp_path)
         snap = AbiSnapshot(library="libfoo.so.1", version="1.0")
-        monkeypatch.setattr("abicheck.service.load_snapshot", lambda _: snap)
+        monkeypatch.setattr("abicheck.workflows.input_resolution.load_snapshot", lambda _: snap)
 
         def _raise(*_a, **_kw):
             raise ScopeMismatchError("scope drift")
@@ -156,7 +156,7 @@ class TestNotComparableExitCode:
     def test_junit_format_emits_errored_testcase(self, tmp_path, monkeypatch):
         old_p, new_p = _write_placeholder_inputs(tmp_path)
         snap = AbiSnapshot(library="libfoo.so.1", version="1.0")
-        monkeypatch.setattr("abicheck.service.load_snapshot", lambda _: snap)
+        monkeypatch.setattr("abicheck.workflows.input_resolution.load_snapshot", lambda _: snap)
 
         def _raise(*_a, **_kw):
             raise ProfileMismatchError("dep.h changed")
@@ -183,7 +183,7 @@ class TestNotComparableExitCode:
         this same function, per the G32 plan's own acceptance criteria)."""
         old_p, new_p = _write_placeholder_inputs(tmp_path)
         snap = AbiSnapshot(library="libfoo.so.1", version="1.0")
-        monkeypatch.setattr("abicheck.service.load_snapshot", lambda _: snap)
+        monkeypatch.setattr("abicheck.workflows.input_resolution.load_snapshot", lambda _: snap)
 
         captured: dict[str, object] = {}
